auto-improve: [#799 Step 1/6] Split `cai_lib/fsm.py` into focused submodules

Refs damien-robotsix/robotsix-cai#818
